% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{build_glmnet_model}
\alias{build_glmnet_model}
\title{Title}
\usage{
build_glmnet_model(data_df, output_name, output_fs_name, loglikelihood_name,
  data_file_name, train_index, intermediate_dir, output_dir)
}
\arguments{
\item{output_dir}{}
}
\description{
Title
}
