% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling_main.R
\name{build_global_model_using_claims_pharma_and_demographics}
\alias{build_global_model_using_claims_pharma_and_demographics}
\title{Build Global model using claims, pharmacy claims and demographics data}
\usage{
build_global_model_using_claims_pharma_and_demographics(train_files, test_files,
  client_codes, intermediate_dir, global_output_dir)
}
\arguments{
\item{train_files}{A vector containing the absolute paths to each of the
training data sets for each client}

\item{test_files}{A vector containing the absolute paths to each of the
testing datasets for each client}

\item{client_codes}{the client codes of each client in the same order as that
of the files in the above vectors}

\item{intermediate_dir}{the path to the directory in which the intermediate
files need to be stored}

\item{global_output_dir}{the output directory to which the model results need to be
stored}
}
\value{
This function does not return anything. It stores the model outputs
in separate files
}
\description{
Build Global model using claims, pharmacy claims and demographics data
}
\examples{
build_global_model_using_claims_pharma_and_demographics(
c("thd_train.csv", "aa_train.csv", "df_train.csv", "pepsi_train.csv"),
c("thd_test.csv", "aa_test.csv", "df_test.csv", "pepsi_test.csv"),
c(1004, 1013, 1011, 1030),
"/home/avinash/intermediate_dir")
}
