% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling_main.R
\name{build_global_model_with_feature_selection}
\alias{build_global_model_with_feature_selection}
\title{Build global model with Feature Selection}
\usage{
build_global_model_with_feature_selection(train_files, test_files, client_codes,
  intermediate_dir, global_output_dir, loglikelihood_ref_file_name)
}
\arguments{
\item{train_files}{the vector containing paths to each of the training files}

\item{test_files}{the vector containing paths to each of the test files}

\item{client_codes}{the client codes vector}

\item{intermediate_dir}{the path to the directory in which the intermediate files and
reference files are stored}

\item{global_output_dir}{the path to the directory in which the output files are stored}

\item{loglikelihood_ref_file_name}{the name of the log likelihood reference file which
should be present in the intermediate directory}
}
\value{
This function does not return anything. It builds the model and stores the
model output to the output directory
}
\description{
Build global model with Feature Selection
}
\examples{
build_global_model_with_feature_selection(train_files, test_files, client_codes,
intermediate_dir, global_output_dir, "ref_file.csv")
}
