% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling_main.R
\name{build_global_model_with_pca}
\alias{build_global_model_with_pca}
\title{Build global model with PCA}
\usage{
build_global_model_with_pca(train_files, test_files, client_codes,
  intermediate_dir, global_output_dir, pca_ref_file)
}
\arguments{
\item{train_files}{the vector containing the paths to the training files}

\item{test_files}{the vector containing the paths to the test files}

\item{client_codes}{the client codes vector}

\item{intermediate_dir}{the path to the directory to which the intermediate files and
reference files are stored}

\item{global_output_dir}{the path to the directory to which the output is to be stored}

\item{pca_ref_file}{the name of the PCA reference file which is stored in the
intermediate directory}
}
\value{
This function does not return anything. It stores the model output in the
output directory
}
\description{
Build global model using PCA and GLMNET
}
\examples{
build_global_model_with_pca(train_files, test_files, client_codes, intermediate_dir,
global_output_dir, pca_ref_file)
}
