% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling_main.R
\name{build_global_model_with_pca_feature_selection}
\alias{build_global_model_with_pca_feature_selection}
\title{Build model with PCA and GLMNET after feature selection}
\usage{
build_global_model_with_pca_feature_selection(train_files, test_files,
  client_codes, intermediate_dir, global_output_dir, pca_ref_file)
}
\arguments{
\item{train_files}{the vector containing the path to the train files}

\item{test_files}{the vector containing the path to the test files}

\item{client_codes}{the vector of client codes in the same order as train files}

\item{intermediate_dir}{the directory in which the intermediate files are to be stored
and the reference files are present}

\item{global_output_dir}{the directory to which the model and pca output are to be
stored}

\item{pca_ref_file}{the name of the pca reference file located in the intermediate
directory}
}
\value{
This function does not return anything. It builds the model and stores the
model output in output_dir
}
\description{
Build model with PCA and GLMNET after feature selection
}
\examples{
build_global_model_with_pca_feature_selection(train_files, test_files, client_codes,
intermediate_dir, global_output_dir, pca_ref_file)
}
