% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling_main.R
\name{build_model_with_claims_and_pharma}
\alias{build_model_with_claims_and_pharma}
\title{Build model with claims, pharmacy and demographics data}
\usage{
build_model_with_claims_and_pharma(data_dir, intermediate_dir, output_dir,
  properties_file, eligibility_file, census_file, census_pca_file = NULL,
  ccs_cpt_file, ccs_bodysystem_file, ccs_icd_file, claims_file, cpt_codes_file,
  pharma_file, threshold = 10, lowerbound = 0, upperbound = 30)
}
\arguments{
\item{data_dir}{The path to the directory in which the data files are present
. Give the complete path to the directory irrespective of the working
directory.}

\item{intermediate_dir}{The path to the directory to which the intermediate
data files are to be stored. Give the complete path to the directory
irrespective of the working directory.}

\item{output_dir}{The path to the directory in which the output files need
to be stored. Give the complete path irrespective of the working directory.}

\item{properties_file}{the name of the properties file}

\item{eligibility_file}{the name of the eligibility file}

\item{census_file}{the name of the census file}

\item{census_pca_file}{the name of the census pca file}

\item{ccs_cpt_file}{the name of the ccs cpt file}

\item{ccs_bodysystem_file}{the name of ccs bodysystem file}

\item{ccs_icd_file}{the name of the ccs icd file}

\item{claims_file}{the name of the claims file}

\item{cpt_codes_file}{the name of the cpt codes file}

\item{pharma_file}{the name of the pharmacy data file}

\item{threshold}{the threshold for grouping claims}

\item{lowerbound}{the lower bound of the readmission interval}

\item{upperbound}{the upper bound of the readmission interval}
}
\value{
This function does not return anything. All the models are built and
 and stored in separate files
}
\description{
Use this function to build model when claims, pharmacy and demographics
data are present
}
\examples{
build_model_with_claims_and_pharma("/home/avinash",
"/home/avinash/intermediate",
"/output/", "properties.csv", "elig.rds", "census.csv", "ccs_cpt.rds",
"ccs_icd.rds", "claims.csv", "cpt_codes.csv", "pharma.csv")

build_model_with_claims("/home/avinash", "/home/avinash/intermediate",
"/output/", "properties.csv", "elig.rds", "census.csv", "census_pca.csv",
"ccs_cpt.rds", "ccs_icd.rds", "claims.csv", "cpt_codes.csv", "pharma.csv")
}
