% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{check_if_readmission}
\alias{check_if_readmission}
\title{Check if readmission}
\usage{
check_if_readmission(data_df, lowerbound, upperbound)
}
\arguments{
\item{data_df}{The data frame which consists of admissions at UPIN level}

\item{upperbound}{}
}
\value{
The data frame with a new column stating readmission or not.
}
\description{
The function checks if there exists a readmission for any admission. If a
readmission exists that readmission is removed.
}
\examples{
check_if_readmssion(admission_df, 0, 30)
}
