% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{consider_non_admission_claims}
\alias{consider_non_admission_claims}
\title{Consider Non Admissions related claims}
\usage{
consider_non_admission_claims(data_df, complete_claims_df, threshold)
}
\arguments{
\item{data_df}{the inpatient claims data frame}

\item{complete_claims_df}{the complete claims data frame}

\item{threshold}{the threshold for grouping}
}
\value{
Admission claims including Non Admission related claims as a dataframe
}
\description{
This function groups the claims after the admission end date which do not have CPT
codes related to admission.
}
\examples{
consider_non_admission_claims(data_df, complete_claims_df, 10)
}
