% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{create_dummy_columns}
\alias{create_dummy_columns}
\title{Create dummy columns from a column which consists of strings separated by a character}
\usage{
create_dummy_columns(data_df, column_name, separator, type, fill,
  remove_original = TRUE)
}
\arguments{
\item{data_df}{the dataframe whose columns are to be spread}

\item{column_name}{tbe column name whose column is to be spread}

\item{separator}{the separator by which the column is concatenated}

\item{type}{the type of the variables which are concatenated to string}

\item{fill}{value to be filled incase the value is missing in the column}

\item{remove_original}{remove the original column or not}
}
\value{
the dataframe with a set of new columns based on the spread
}
\description{
Create dummy columns from a column which consists of strings separated by a character
}
\examples{
create_dummy_columns(admission_df, "admitting_diagnosis_bs", "_", "numeric", 0, TRUE)
}
