% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{do_feature_selection_by_loglikelihood}
\alias{do_feature_selection_by_loglikelihood}
\title{Title}
\usage{
do_feature_selection_by_loglikelihood(directory, ref_file, file_name,
  output_name)
}
\arguments{
\item{output_name}{}
}
\description{
Title
}
