% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{find_admitting_diagnosis_admission_level}
\alias{find_admitting_diagnosis_admission_level}
\title{Find admitting diagnosis at admission level}
\usage{
find_admitting_diagnosis_admission_level(data_df, complete_claims_df,
  bodysystem_df, ccs_df)
}
\arguments{
\item{data_df}{the data frame which contains the admission}

\item{complete_claims_df}{the complete claims data frame}

\item{bodysystem_df}{the bodysystem data frame}

\item{ccs_df}{the CCS ICD dataframe}
}
\value{
the data frame with new columns for primary diagnoses at both bodysystem level
and CCS level
}
\description{
Find admitting diagnosis at admission level
}
\examples{
find_admittin_diagnosis_admission_level(data_df, complete_claims_df, bodysystem_df,
ccs_df)
}
