% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{find_blood_transfusion}
\alias{find_blood_transfusion}
\title{Find blood transfusion}
\usage{
find_blood_transfusion(data_df, blood_transfusion_claims)
}
\arguments{
\item{data_df}{- the data frame with consists of admissions}

\item{blood_transfusion_claims}{- the data frame which consists of blood
transfusion claims}
}
\value{
This function returns the data frame with a blood transfusion boolean
column
}
\description{
Find blood transfusion
}
\examples{
find_blood_transfusion(admissions_df, blood_transfusion_claims_df)
}
