% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_admission_medications}
\alias{get_admission_medications}
\title{Get admission medications}
\usage{
get_admission_medications(admission_df, pharma_df)
}
\arguments{
\item{admission_df}{the admission dataframe}

\item{pharma_df}{the pharmacy claims dataframe at gpi level}
}
\value{
the data frame containg admissions and medications
}
\description{
A function to find the medications a person is under during the admission
}
\examples{
get_admission_medications(admission_df, pharma_df)
}
