% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_admission_procedures}
\alias{get_admission_procedures}
\title{Get Admission Procedures}
\usage{
get_admission_procedures(admission_with_admit_diagnosis_df, complete_claims_df,
  ccs_cpt_file)
}
\arguments{
\item{admission_with_admit_diagnosis_df}{The dataframe with admissions}

\item{complete_claims_df}{the complete claims data frame}

\item{ccs_cpt_file}{the CCS CPT File}
}
\value{
the data frame with admission procedures at CCS CPT level
}
\description{
This function is used to find the procedures underwent by a patient during an admission
}
\examples{
get_admission_procedures(admission_df, complete_claims_df, "ccs_cpt.csv")
}
