% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_admitting_diagnosis}
\alias{get_admitting_diagnosis}
\title{Get Admitting Diagnosis}
\usage{
get_admitting_diagnosis(inpatient_claims_readmission, complete_claims_df,
  ccs_bodysystem_file, ccs_icd_file, data_dir)
}
\arguments{
\item{inpatient_claims_readmission}{the admission dataframe}

\item{complete_claims_df}{the complete claims dataframe}

\item{ccs_bodysystem_file}{the ccs to bodysystem mapping file}

\item{ccs_icd_file}{the ccs to icd mapping file}
}
\value{
the admission data frame that consists primary diagnoses at bodysystem level
}
\description{
This function will get all the unique primary diagnoses at bodysystem level during the
admission
}
\examples{
get_admitting_diagnosis(inpatient_claims_readmission, complete_claims_df, "ccs_bs.csv",
"ccs_icd.csv")
}
