% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_comorbidities}
\alias{get_comorbidities}
\title{Get comorbidities}
\usage{
get_comorbidities(data_df, complete_claims_df)
}
\arguments{
\item{data_df}{the data frame containing te admissions and diagnosis codes
in each admission}
}
\value{
- the comorbidites during each admission
}
\description{
This function is used to find the comorbidities present during the time of
admissiong
}
\examples{
get_comorbidities(admission_df)
}
