% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_cpt_claims_grouped}
\alias{get_cpt_claims_grouped}
\title{Get emergency claims}
\usage{
get_cpt_claims_grouped(claims_df, cpt_v, threshold = 10)
}
\arguments{
\item{claims_df}{the claims data frame}

\item{cpt_codes_df}{the cpt codes data frame}
}
\value{
- the data frame containing emergency admissions
}
\description{
A function to filter emergency claims to find emergency admissions.
}
\examples{
get_cpt_claims_grouped(claims_df, cpt_codes_df)
}
