% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_icd_comorbidities}
\alias{get_icd_comorbidities}
\title{Get ICD comorbidities}
\usage{
get_icd_comorbidities(admission_df, complete_claims_df)
}
\arguments{
\item{admission_df}{the admission data frame}

\item{inpatient_claims_df}{the claims related to admissions}
}
\value{
data frame consisting of admissions and comorbidities
}
\description{
Function to find the comorbidities present at the time of admission.
}
\examples{
get_icd_comorbidities(admission_df, claims_df)
}
