% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_length_of_stay}
\alias{get_length_of_stay}
\title{get_length_of_stay}
\usage{
get_length_of_stay(data_df)
}
\arguments{
\item{data_df}{- the data frame consisting of admissions}
}
\value{
- the dataframe with a new column length_of_stay
}
\description{
Given an admission dataframe this function finds the length of stay at
hospital
}
\examples{
get_length_of_stay(admissions_df)
}
