% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_new_claim}
\alias{get_new_claim}
\title{Get New Claim}
\usage{
get_new_claim(new_claim, first_claim, second_claim)
}
\arguments{
\item{new_claim}{The claim whose service start date and service end date need to be
changed}

\item{first_claim}{The first claim}

\item{second_claim}{The second claim}
}
\value{
New claim with changed service start and end dates
}
\description{
Given two claims this function will merge the two claims and get the min and max
service start date and return the new claim.
}
\examples{
get_new_claim(new_claim, first_claim, second_claim)
}
