% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_pcs}
\alias{get_pcs}
\title{Get Number of Principal Components for 90% variance}
\usage{
get_pcs(pca_output, output_dir, variance = 0.9)
}
\arguments{
\item{pca_output}{The output of PCA}

\item{output_dir}{the output dir to which the scree plots need to be stored}
}
\value{
the number of principal components to be considered
}
\description{
Get Number of Principal Components for 90% variance
}
\examples{
get_pcs(pca_output, "/home/avinash/output_dir", 0.85)

get_pcs(pca_output, "/home/avinash/output_dir")
}
