% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_previous_admissions}
\alias{get_previous_admissions}
\title{Get previous admissions}
\usage{
get_previous_admissions(data_df)
}
\arguments{
\item{data_df}{the data frame consisting of admissions}
}
\value{
the data frame with a column containing #previous admissions
}
\description{
This function finds the number of previous admissions before each admission.
}
\examples{
get_previous_admission(admission_df)
}
