% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_previous_admissions_for_each_admission}
\alias{get_previous_admissions_for_each_admission}
\title{Get previous admissions for each admission}
\usage{
get_previous_admissions_for_each_admission(data_df)
}
\arguments{
\item{data_df}{the data frame which consists of admissions}
}
\value{
data frame with new column as the number of previous readmissions
}
\description{
This function gets the number of admissions before each admissions at UPIN
level.
}
\examples{
get_previous_admissions_for_each_admission(admission_df)
}
