% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_previous_emergency_admissions}
\alias{get_previous_emergency_admissions}
\title{Get previous emergency admissions}
\usage{
get_previous_emergency_admissions(inpatient_claims_readmission,
  emergency_claims_df)
}
\arguments{
\item{inpatient_claims_readmission}{the dataframe containing the admissions}

\item{emergency_claims_df}{the claims related to emergency admissions}
}
\value{
a dataframe with the number of previous emergency admissions
}
\description{
A function that finds the number of emergency admissions before each
admission.
}
\examples{
get_previous_emergency_admissions(claims_df, emergency_df)
}
