% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_previous_hospital_visits}
\alias{get_previous_hospital_visits}
\title{Get Previous hospital visits}
\usage{
get_previous_hospital_visits(inpatient_claims_readmission,
  initial_observation_claims)
}
\arguments{
\item{initial_observation_claims}{the claims related to hospital visits.}

\item{inpatient_claims_with_admit_procedures}{the inpatient claims df which consists of
admissions}
}
\value{
the admission dataframe with a new column as the number of previous hospital
visits
}
\description{
This function gets the number of previous hospital visits by a person before each
admission
}
\examples{
get_previous_hospital_visits(inpatient_claims_df, visit_claims_df)
}
