% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_procedures_at_admission_level}
\alias{get_procedures_at_admission_level}
\title{Get Procedures at Admission level}
\usage{
get_procedures_at_admission_level(data_df, complete_claims_df, cpt_ccs_df)
}
\arguments{
\item{data_df}{the data frame of the admission}

\item{complete_claims_df}{the complete claims dataframe}

\item{cpt_ccs_df}{the CPT CCS dataframe}
}
\value{
The admission dataframe with procedures during the admision.
}
\description{
This function takes each admission and finds the procedures during the admission.
}
\examples{
get_procedures_at_admission_level(data_df, complete_claims_df, cpt_ccs_df)
}
