% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{get_properties_data}
\alias{get_properties_data}
\title{get_properties_data}
\usage{
get_properties_data(data_dir, properties_file)
}
\arguments{
\item{data_dir}{the path to the directory in which the file exists}

\item{properties_file}{the name of the properties file}
}
\value{
the properties file as a data frame
}
\description{
This function reads the properties file and returns the data frame
}
\examples{
get_properties_data("/home/avinash", "properties.csv")

get_properties_data("/home/avinash", "properties.rds")

get_properties_data("/home/avinash", "properties.tsv")
}
