% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{group_additional_claims}
\alias{group_additional_claims}
\title{Group Additional Claims}
\usage{
group_additional_claims(data_df, to_group_df, threshold)
}
\arguments{
\item{data_df}{the already present admission dataframe}

\item{to_group_df}{the claims dataframe which should be grouped}

\item{threshold}{the threshold for grouping}
}
\value{
the dataframe after grouping additional claims with the admission dataframe
}
\description{
Group additional claims for each admission based on the claims dataframe that is passed
}
\examples{
group_additional_claims(data_df, to_group_df, 10)
}
