% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{group_claims}
\alias{group_claims}
\title{Group claims}
\usage{
group_claims(data_df, threshold, recent = FALSE)
}
\arguments{
\item{data_df}{the data frame which is to be grouped}
}
\value{
the claims grouped at admission level.
}
\description{
This function is a wrapper to group consecutive claims at UPIN and admission
level.
}
\examples{
group_claims(claims_df)
}
