% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{group_consecutive_claims}
\alias{group_consecutive_claims}
\title{Group consecutive claims}
\usage{
group_consecutive_claims(data_df, threshold, recent = FALSE)
}
\arguments{
\item{data_df}{the dataframe which is to be grouped}

\item{threshold}{the number of days before which a claims can be
considered as a part of the same admission}

\item{recent}{if TRUE, then as soon as a claim exceeds the threshold the
grouping stops}
}
\value{
- the grouped claims data
}
\description{
The function groups claims based on their service start dates. Two claims are
grouped together if their service start date difference <= 10 days
}
\examples{
group_consecutive_claims(claims_df, 15)

group_consecutive_claims(claims_df, 20)

group_consecutive_claims(claims_df, 2, TRUE)
}
