% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_main.R
\name{predict_main_pca_glmnet}
\alias{predict_main_pca_glmnet}
\title{Title}
\usage{
predict_main_pca_glmnet(data_dir, intermediate_dir, output_dir, properties_file,
  eligibility_file, census_file, census_pca_file = NULL, ccs_cpt_file,
  ccs_bodysystem_file, ccs_icd_file, claims_file, cpt_codes_file, pharma_file,
  threshold = 10, lowerbound = 0, upperbound = 30, model_file,
  pca_output_file, number_of_principal_components_file, reference_file)
}
\arguments{
\item{reference_file}{}
}
\description{
Title
}
