% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_functions.R
\name{prediction_with_glmnet}
\alias{prediction_with_glmnet}
\title{Title}
\usage{
prediction_with_glmnet(model_file, test_data_file, output_dir, reference_file)
}
\arguments{
\item{reference_file}{}
}
\description{
Title
}
