% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_functions.R
\name{prediction_with_pca_glmnet}
\alias{prediction_with_pca_glmnet}
\title{Title}
\usage{
prediction_with_pca_glmnet(model_file, test_data_file, output_dir,
  pca_output_file, number_of_principal_components_file, reference_file)
}
\arguments{
\item{reference_file}{}
}
\description{
Title
}
