% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_data.R
\name{preprocess_claims}
\alias{preprocess_claims}
\title{Pre process claims}
\usage{
preprocess_claims(data_dir, intermediate_dir, claims_file, props_df,
  cpt_codes_file, ccs_cpt_file, ccs_bodysystem_file, ccs_icd_file, threshold,
  lowerbound, upperbound)
}
\arguments{
\item{data_dir}{The path to the directory in which the data is present}

\item{intermediate_dir}{the path to the directory in which the intermediate
files need to be stored}

\item{claims_file}{the name of the claims file}

\item{props_df}{the properties dataframe}

\item{cpt_codes_file}{the name of the cpt codes file}

\item{ccs_cpt_file}{the name of the ccs cpt file}

\item{ccs_bodysystem_file}{the name of the icd to bodysystem file}

\item{ccs_icd_file}{the name of the ccs icd file}

\item{threshold}{the threshold for grouping claims}

\item{lowerbound}{the lower bound of readmisison interval}

\item{upperbound}{the upper bound of readmission interval}
}
\value{
the admission dataframe
}
\description{
This function reads the claims file and finds the admissions, readmissions,
emergency admissions, procedures, comorbidities.
}
\examples{
preprocess_claims('/home', '/home/intermediate', 'claims.csv', props_df,
'cpt_codes.csv')
}
