% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{process_census}
\alias{process_census}
\title{process_census}
\usage{
process_census(data_dir, census_file, props_df, elig_df)
}
\arguments{
\item{data_dir}{the path to the directory in which the file exists}

\item{props_df}{the properties data frame}

\item{elig_file}{the name of the census file}
}
\value{
the census file as a data frame
}
\description{
This function reads the census file and returns the dataframe
}
\examples{
process_census("/home/avinash", "census.csv", properties_df)
}
