% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{process_census_pca}
\alias{process_census_pca}
\title{process_census_pca}
\usage{
process_census_pca(data_dir, census_pca_file)
}
\arguments{
\item{data_dir}{the path to the directory in which the file is present}

\item{census_pca_file}{the name of the census pca file}
}
\value{
the census pca file as a dataframe
}
\description{
The function reads the census pca file and returs it as a data frame
}
\examples{
process_census_pca("/home/avinash", "census_pca.csv")
}
