% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{process_claims}
\alias{process_claims}
\title{process_claims}
\usage{
process_claims(data_dir, claims_file, props_df, filter_criteria = NULL)
}
\arguments{
\item{data_dir}{the path to the directory in which the claims are present}

\item{claims_file}{the pattern of the claims file}

\item{props_df}{the properties df}

\item{filter_criteria}{the filter criteria based on which the claims will be
filtered}
}
\value{
the claims file filtered by cpt codes as a dataframe
}
\description{
The function reads claims data using read_claims_data in EviveUtil and
returns them as a dataframe
}
\examples{
process_claims("/home/avinash", "_claims.csv", props_df, cpt_codes_df)
}
