% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{process_cpt_codes}
\alias{process_cpt_codes}
\title{process_cpt_codes}
\usage{
process_cpt_codes(data_dir, cpt_codes_file)
}
\arguments{
\item{data_dir}{the path of the directory in which the file exists}

\item{cpt_codes_file}{the name of the cpt codes file}
}
\value{
- the cpt codes file as a data frame
}
\description{
This function reads the cpt codes from the file and returns it as a dataframe
}
\examples{
process_cpt_codes("/home/avinash", "cpt_codes.csv")
}
