% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{process_elig}
\alias{process_elig}
\title{process_elig}
\usage{
process_elig(data_dir, elig_file, props_df)
}
\arguments{
\item{data_dir}{the path to the directory in which the file exists}

\item{elig_file}{the name of the eligibility file}

\item{props_df}{the properties data frame}
}
\value{
the eligibility file as a data frame
}
\description{
This function reads the eligibility file and returns the dataframe
}
\examples{
process_elig("/home/avinash", "eligibility.csv", properties_df)
}
